pyproject.toml
[project]
name = "finetuning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agent-framework>=1.0.0rc2",
    "datasets==3.6.0",
    "google-cloud-storage>=3.9.0",
    "google-genai>=1.63.0",
    "ipykernel>=7.2.0",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.24.0",
    "pandas>=2.3.3",
    "pydantic-settings>=2.13.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rapidfuzz>=3.14.3",
    "tiktoken>=0.12.0",
    "transformers>=5.2.0",
]
